Atlas-standalone: Query optimisation on DAO for expression analytics - fetch by experiment instead of doing a query per gene.  Added atlas count info to index builder.  Ready for a test run.
git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@9446 2913f559-6b04-0410-9a09-c530ee9f5186

+++ indexbuilder/src/main/java/uk/ac/ebi/ae3/indexbuilder/service/GeneAtlasIndexBuilderService.java
@@ -172,7 +172,7 @@ public class GeneAtlasIndexBuilderService extends IndexBuilderService {
         getAtlasDAO().getExpressionAnalyticsByGeneID(geneID);
 
     for (ExpressionAnalysis expressionAnalytic : expressionAnalytics) {
-      Long experimentId = expressionAnalytic.getExperimentID();
+      Long experimentId = (long)expressionAnalytic.getExperimentID();
       if (experimentId == 0) {
         log.error("Found experimentId=0 for gene " + geneID);
         continue;
